+++ sdk/cosmos/azure-cosmos/CHANGELOG.md
@@ -15,6 +15,7 @@
 #### Other Changes
 * Changed to include client id in headers for all requests. See [PR 42104](https://github.com/Azure/azure-sdk-for-python/pull/42104).
 * Added an option to override AAD audience scope through environment variable. See [PR 42228](https://github.com/Azure/azure-sdk-for-python/pull/42228).
+* Added session token false progress merge logic. See []()
 
 ### 4.14.0b1 (2025-07-14)